% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_labels.R
\name{all_labels}
\alias{all_labels}
\title{A list with all labels available for all occupational schemas}
\format{
list
}
\description{
Each slot contains a data frame for each of the schemas. The name of each
slot in the list shows the current schema.
}
