% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_schemas.R
\name{all_schemas}
\alias{all_schemas}
\title{A list with all available translations between occupational schemas.}
\format{
list
}
\description{
Each slot contains a data frame for each of the translation between two schemas.
Additionally, it contains the translation between major/submajor/minor/unit for
all ISCO schemas. The name of each slot in the list shows which schemas being
translated. Always, the first column is the origin schema and the second column
is the destination schema.
}
