% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esco.R
\name{esco_to_isco08}
\alias{esco_to_isco08}
\title{Translate 4-digit ESCO to ISCO08}
\usage{
esco_to_isco08(x, label = FALSE, to_factor = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ESCO codes.}

\item{label}{A logical value indicating whether to return the labels of the translated ISCO08 codes (default is \code{FALSE}).}

\item{to_factor}{A logical value indicating whether to return a factor instead of a character. The order of the labels is taken from the labels for ISCO08 found in \code{all_labels} (default is \code{FALSE}).}
}
\value{
A character vector of ISCO08 codes.
}
\description{
This function translates a vector of 4-digit ESCO codes to ISCO08 codes using the translation table stored in \code{all_schema$esco_to_isco08}.
}
\details{
This translation converts the ESCO classification of the European Commission to ISCO08. This translations uses ESCO v1.1.1 and was downloaded from the dataset \url{https://esco.ec.europa.eu/en/use-esco/download}.

For more information on this class schema, please check the references below:
\itemize{
\item The ESCO Classification: \url{https://esco.ec.europa.eu/en/classification}
}
}
\examples{

esco_to_isco08(c("11401", "1101", "11301"), label = FALSE)

esco_to_isco08(c("11401", "1101", "11301"), label = TRUE)

esco_to_isco08(c("11401", "1101", "11301"), label = TRUE, to_factor = TRUE)

}
