% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess}
\alias{ess}
\title{An example data frame with data from the European Social Survey, round 5.}
\format{
dataframe
}
\description{
Just an example data frame to show how the package can be used to translate
between schemas.
}
