% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R, R/isco88.R
\name{isco08_to_esec}
\alias{isco08_to_esec}
\alias{isco88com_to_esec}
\title{Translate 3-digit ISCO08/ISCO88COM to ESEC}
\usage{
isco08_to_esec(x, is_supervisor, self_employed, n_employees, label = FALSE)

isco88com_to_esec(
  x,
  is_supervisor,
  self_employed,
  n_employees,
  full_method = TRUE,
  label = FALSE
)
}
\arguments{
\item{x}{A character vector of 3-digit ISCO codes. This should be the 4-digit equivalent so instead of 131, the code should be 1310, which is the 4-digit version of of the 3-digit ISCO.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or an employee (0).}

\item{n_employees}{A numeric vector indicating the number of employees under each respondent.}

\item{label}{A logical value indicating whether to return the labels of the translated ESEC codes (default is \code{FALSE}).}

\item{full_method}{a boolean on whether to apply the full method or the simple method.}
}
\value{
A character vector of ESEC codes.
}
\description{
This function translates a vector of 3-digit ISCO08/ISCO08COM codes to ESEC codes using the translation tables stored in \code{all_schema$isco08_to_esec / all_schema$isco88com_to_esec_three}.
}
\details{
This translation was taken from the \code{iscogen} Stata package. For more details, check out the package documentation and search for \code{ISCO08/ISCO88COM -> ESEC}.

This function will accept 3 digit codes as 4 digits. This means that if the 3-digit code is 131 then it should be 1310. All codes should be 4 digits, even though the code is represented as 3-digits (1310, 1320, etc..)

ISCO88COM has two types of translations: simple and full method. The full method uses information on whether the respondent is a supervisor, self-employed and the number of subordinates of the employee. In contrast, the simple method matches directly the ISCO code to an ESEC code.

For more info, please see page 17 of the European Socio-economic Classification (ESeC) User Guide (2006) by Rode, D. and Harrison, E.

The translation for ISCO88 is done from ISCO88COM which is not ISCO88. If you have ISCO88, you can translate it to ISCO88COM using the function \code{DIGCLASS::isco88_to_isco88com} before translating to ESEC.

Contrary to ISCO88COM-ESEC, ISCO08 does not have a simplified method and the translation is done from ISCO08 directly to ESEC.
}
\examples{
library(dplyr)

# convert isco08 to three digits
ess$isco08_three <- isco08_swap(ess$isco08, from = 4, to = 3)

ess \%>\%
  transmute(
    isco08_three,
    esec = isco08_to_esec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    ),
    esec_label = isco08_to_esec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    )
  )

# convert isco88 to three digits
ess$isco88com_three <- isco88_swap(ess$isco88com, from = 4, to = 3)

# Using the full method
ess \%>\%
  transmute(
    isco88com_three,
    esec_label = isco88com_to_esec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE,
      full_method = TRUE
    ),
    esec_no_label = isco88com_to_esec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE,
      full_method = TRUE
    )
  )

# Using the simple method
ess \%>\%
  transmute(
    isco88com_three,
    esec_simple = isco88com_to_esec(
      isco88com_three,
      label = FALSE,
      full_method = FALSE
    ),
    esec_simple_label = isco88com_to_esec(
      isco88com_three,
      label = TRUE,
      full_method = FALSE
    )
  )

}
