% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R
\name{isco08_to_eseg}
\alias{isco08_to_eseg}
\title{Translates 2-digit ISCO08 to ESEG.}
\usage{
isco08_to_eseg(
  x,
  work_status,
  main_activity,
  age,
  type,
  label = FALSE,
  to_factor = FALSE
)
}
\arguments{
\item{x}{A character vector of 2-digit ISCO codes. This should be the 4-digit equivalent so instead of 13, the code should be 1300, which is the 4-digit version of of the 2-digit ISCO.}

\item{work_status}{A numeric vector of values from 0 to 3 where \code{1 = self_employed}, \code{0 = employee} and \verb{2 = non employed}.}

\item{main_activity}{A numeric vector of values from 1 to 5 where \verb{1 = respondent is working}, \verb{2 = respondent is in education}, \verb{3 = respondent is disabled }, \verb{4 = respondent has no paid work (household work, taking care of children, etc..)} and \verb{5 = respondent is retired}. For an example, see the variable \code{mainact} from the European Social Survey.}

\item{age}{A numeric vector of ages of the respondent.}

\item{type}{The type of translation to make. Possible values are "one-digit" and "two-digit". The "one-digit" translation returns a broad summary based translation of only 9 categories, whereas the "two-digit" translation returns a much bigger ESEG translation of more than 25 categories.}

\item{label}{A logical value indicating whether to return the labels of the translated ESEG codes (default is \code{FALSE}).}

\item{to_factor}{A logical value indicating whether to return a factor instead of a character. The order of the labels is taken from the sorted codes of ESEG which can be found in the source code of each function.}
}
\description{
This function translates a vector of 2-digit ISCO88COM codes to the ESEG class schema.
}
\details{
The translation was implemented following the work of Kea Tijdens in the document "ESEG-2014 coding scheme + explanatory note". For more details, see the last table in the document \href{https://seriss.eu/wp-content/uploads/2016/12/SERISS-Deliverable-D8-13-ESeG-coding_submitted.pdf}{here}.

For more details on this class schema, please check the references below:
\itemize{
\item Börlin, S., & Zettl, L. (2020). Mikrozensus Tools: Die Klassifikation European Socio-economic Groups (ESeG) im Mikrozensus Scientific Use File. (GESIS Papers, 2020/08). Köln: GESIS - Leibniz-Institut für Sozialwissenschaften. \url{https://doi.org/10.21241/ssoar.68449}
\item Bohr, J. (2018). EU-AES Tools: Implementation of the European Socioeconomic Groups Classification (ESeG) using Adult Education Survey Microdata. (GESIS Papers, 2018/14). Köln: GESIS - Leibniz-Institut für Sozialwissenschaften.\url{https://doi.org/10.21241/ssoar.57622}
\item Meron M, and all ESSnet members (2014) ESSnet ESeG Final Report. Paris, INSEE,Direction des Statistiques Démographiques et Sociales ESSnet project
\item Meron, M. et al. (2014): Final Report of the ESSnet on the harmonisation and implementation of a European socio-economic classification: European Socio-economic Groups (ESeG)
\item Tijdens, K.G. (2016) ESEG-2014 coding scheme + explanatory note. Deliverable 8.13 of the SERISS project funded under the European Union’s Horizon 2020 research and innovation programme GA No: 654221. Available at: \href{www.seriss.eu/resources/deliverables}{https://seriss.eu/resources/deliverables}
}

Resource websites of the European Socio-economic Groups (ESeG):
\itemize{
\item \url{https://cros-legacy.ec.europa.eu/content/eseg_en}
\item \url{https://ec.europa.eu/eurostat/cros/content/eseg-report-technicaldocuments_en}
\item \url{https://ec.europa.eu/eurostat/cros/system/files/ESEG-Report-TechnicalAnnexes_0.zip_en}
}
}
\examples{
library(dplyr)

# Convert isco08 to two digits
ess$isco08_two <- isco08_swap(ess$isco08, from = 4, to = 2)

# Using the two-digit translation
ess \%>\%
  transmute(
    isco08_two,
    eseg = isco08_to_eseg(
      isco08_two,
      work_status,
      main_activity,
      agea,
      type = "two-digit"
    ),
    eseg_label = isco08_to_eseg(
      isco08_two,
      work_status,
      main_activity,
      agea,
      type = "two-digit",
      label = TRUE
    )
  )

# Using the one-digit translation
ess \%>\%
  transmute(
    isco08_two,
    eseg = isco08_to_eseg(
      isco08_two,
      work_status,
      main_activity,
      agea,
      type = "one-digit"
    ),
    eseg_label = isco08_to_eseg(
      isco08_two,
      work_status,
      main_activity,
      agea,
      type = "one-digit",
      label = TRUE
    )
  )

}
