% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R, R/isco88.R
\name{isco08_to_ipics}
\alias{isco08_to_ipics}
\alias{isco88_to_ipics}
\title{Translate 4-digit ISCO08/ISCO88 to IPICS}
\usage{
isco08_to_ipics(x, self_employed, n_employees, label = FALSE)

isco88_to_ipics(x, self_employed, n_employees, label = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or an employee (0).}

\item{n_employees}{A numeric vector indicating the number of employees under each respondent.}

\item{label}{A logical value indicating whether to return the labels of the translated IPICS codes (default is \code{FALSE}).}
}
\value{
A character vector of IPICS codes.
}
\description{
This function translates a vector of 4-digit ISCO08/ISCO88 codes to IPICS codes using the translation tables stored in \code{all_schema$isco08_to_ipics / all_schema$isco88_to_ipics}.
}
\details{
These translation were created from the CSV files shared by Oscar Smallenbroek name "ISCO08 to IPICS.csv" and "ISCO88 to IPICS.csv". For more info, please contact the author.

For more details on the class schema please check the references below:
\itemize{
\item Hertel, Florian R. 2017. Social Mobility in the 20th Century: Class Mobility and Occupational Change in the United States and Germany. Springer VS.
}
}
\examples{
library(dplyr)

# isco08
ess \%>\% transmute(
  isco08,
  ipics = isco08_to_ipics(isco08, self_employed, emplno),
  ipics_label = isco08_to_ipics(isco08, self_employed, emplno, label = TRUE)
)


# isco88
ess \%>\% transmute(
  isco88,
  ipics = isco88_to_ipics(isco88, self_employed, emplno),
  ipics_label = isco88_to_ipics(isco88, self_employed, emplno, label = TRUE)
)

}
