% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R
\name{isco08_to_microclass}
\alias{isco08_to_microclass}
\title{Translate 4-digit ISCO08 to microclass}
\usage{
isco08_to_microclass(x, label = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO08 codes.}

\item{label}{A logical value indicating whether to return the labels of the translated microclass codes (default is \code{FALSE}).}
}
\value{
A character vector of microclass codes.
}
\description{
This function translates a vector of 4-digit ISCO08 codes to microclass codes using the translation table stored in \code{all_schema$isco08_to_microclass}.
}
\details{
This translation was created from the Excel file shared by Oscar Smallenbroek called "isco08 to micro with numeric labels.xlsx". For more info, please contact the author.
}
\examples{
library(dplyr)

ess \%>\% transmute(
  isco08,
  microclasses = isco08_to_microclass(isco08),
  microclasses_label = isco08_to_microclass(isco08, label = TRUE)
)

}
