% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R
\name{isco08_to_microclasses}
\alias{isco08_to_microclasses}
\title{Translate 4-digit ISCO08 to Microclasses}
\usage{
isco08_to_microclasses(x)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO08 codes.}
}
\value{
A character vector of Microclasses codes.
}
\description{
This function translates a vector of 4-digit ISCO08 codes to Microclasses codes using the
translation table stored in the \code{all_schemas$isco08_to_microclasses} data frame.
}
\details{
This translation was created from the csv file attached with the document "Creation of the cross walk from ISCO 2008 to micro" written by Oscar Smallenbroek, Florian Hertel and Carlo Barone. For more info, please contact the authors.

This function does not have numeric categories so when translating from ISCO08, labels will be returned. The argument \code{label} does not apply to this function.
}
\examples{
library(dplyr)

# Using the full method
# ess \%>\%
#  transmute(microclasses = isco08_to_microclasses(isco08))

}
