% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R, R/isco88.R
\name{isco08_to_oep}
\alias{isco08_to_oep}
\alias{isco88_to_oep}
\title{Translate ISCO codes to OEP (Occupational Earning Potential)}
\usage{
isco08_to_oep(x, to_factor = FALSE)

isco88_to_oep(x, to_factor = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO08/ISCO88 codes}

\item{to_factor}{A logical value indicating whether to return a factor instead of a character}
}
\value{
A character vector with OEP values
}
\description{
Translates ISCO08/ISCO88 codes to OEP, a hierarchical indicator of occupations' earning potential.
OEP is a numeric scale that measures occupations' median earnings and expresses them as percentiles
of the overall earnings structure.
}
\details{
This function expects 4-digit ISCO codes. For different digit levels (1-3), first convert
using \code{isco08_swap()} or \code{isco88_swap()}. For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# For 3-digit ISCO:
df$isco08_3d <- isco08_swap(df$isco08, from = 4, to = 3)
df$oep <- isco08_to_oep(df$isco08_3d)
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

# Using 4-digit ISCO (default)
ess \%>\%
  transmute(
    isco08,
    oep = isco08_to_oep(isco08)
  )

# Using 3-digit ISCO
ess \%>\%
  transmute(
    isco08,
    isco08_3d = isco08_swap(isco08, from = 4, to = 3),
    oep = isco08_to_oep(isco08_3d)
  )

# Using 2-digit ISCO
ess \%>\%
  transmute(
    isco08,
    isco08_2d = isco08_swap(isco08, from = 4, to = 2),
    oep = isco08_to_oep(isco08_2d)
  )

# Using 1-digit ISCO
ess \%>\%
  transmute(
    isco08,
    isco08_1d = isco08_swap(isco08, from = 4, to = 1),
    oep = isco08_to_oep(isco08_1d)
  )

}
