% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco68.R
\name{isco68_to_egp}
\alias{isco68_to_egp}
\title{Translate ISCO68 to EGP}
\usage{
isco68_to_egp(x, self_employed, n_employees, label = FALSE)
}
\arguments{
\item{x}{A character vector of ISCO68 codes.}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees for each individual.}

\item{label}{A logical value indicating whether to return the labels of the translated EGP codes (default is \code{FALSE}).}
}
\value{
A character vector of EGP codes.
}
\description{
This function translates a vector of ISCO68 codes to EGP codes using the
translation table stored in the \code{all_schemas$isco68_to_egp} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(EGP = isco68_to_egp(isco68, self_employed, emplno, label = TRUE))
ess \%>\% mutate(EGP = isco68_to_egp(isco68, self_employed, emplno, label = FALSE))

}
