% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco68.R
\name{isco68_to_egp_mp}
\alias{isco68_to_egp_mp}
\title{Translate ISCO68 to EGP-MP}
\usage{
isco68_to_egp_mp(x, is_supervisor, self_employed, n_employees, label = FALSE)
}
\arguments{
\item{x}{A character vector of ISCO68 codes.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees for each individual.}

\item{label}{A logical value indicating whether to return the labels of the
translated EGP-MP codes (default is \code{FALSE}).}
}
\value{
A character vector of EGP-MP codes.
}
\description{
This function translates a vector of ISCO68 codes to EGP-MP codes.
}
\details{
EGP-MP is a class schema similar to EGP but reassigns managers and professionals (ISCO68 codes 1 and 2) to have both high/low managers and profesionals. Note that since this translation first converts ISCO68 to EGP and then applies the following logic to build EGP-MP:
\itemize{
\item All occupations with EGP digit 1 and ISCO 1-digit 0 or 1 or has subordinates, \strong{is a high manager}
\item All occupations with EGP digit 1 and is self-employed with more than 1 employee, \strong{is a high manager}
\item All occupations with EGP digit 1 and has a 1-digit ISCO higher than 1 and is either an employee or a self-employed with no subordinates, is a \strong{high professional}
\item All occupations with EGP digit 2 and ISCO 1-digit 0 or 1 or has subordinates, is a \strong{lower manager}
\item All occupations with EGP digit 2 and is self-employed with more than 1 employee, is a \strong{lower manager}
\item All occupations with EGP digit 2 and has a 1-digit ISCO higher than 1 and is either an employee or a self-employed with no subordinates, is a \strong{lower professional}
}

All other EGP codes remain the same.

This translation was created from the Stata do file shared by Oscar Smallenbroek called "ESEC-MP.do". For more info, please contact the author.
}
\examples{
library(dplyr)

ess \%>\%
  transmute(
    isco68,
    egp_mp = isco68_to_egp_mp(
      isco68,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    ),
    egp_mp_label = isco68_to_egp_mp(
      isco68,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    )
  )

}
