% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco68.R
\name{isco68_to_isco08}
\alias{isco68_to_isco08}
\title{Translate ISCO68 to ISCO08}
\usage{
isco68_to_isco08(x, label = FALSE)
}
\arguments{
\item{x}{A character vector of ISCO68 codes.}

\item{label}{A logical value indicating whether to return the labels of the
translated ISCO08 codes (default is \code{FALSE}).}
}
\value{
A character vector of ISCO08 codes.
}
\description{
This function translates a vector of ISCO68 codes to ISCO08 codes using the
translation table stored in the \code{all_schemas$isco68_to_isco08} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(ISCO08 = isco68_to_isco08(isco68, label = TRUE))
ess \%>\% mutate(ISCO08 = isco68_to_isco08(isco68, label = FALSE))

}
