% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco68.R
\name{isco68_to_isco88}
\alias{isco68_to_isco88}
\title{Translate ISCO68 to ISCO88}
\usage{
isco68_to_isco88(x, label = FALSE)
}
\arguments{
\item{x}{A character vector of ISCO68 codes.}

\item{label}{A logical value indicating whether to return the labels of the
translated ISCO88 codes (default is \code{FALSE}).}
}
\value{
A character vector of ISCO88 codes.
}
\description{
This function translates a vector of ISCO68 codes to ISCO88 codes using the
translation table stored in the \code{all_schemas$isco68_to_isco88} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(ISCO88 = isco68_to_isco88(isco68, label = TRUE))
ess \%>\% mutate(ISCO88 = isco68_to_isco88(isco68, label = FALSE))

}
