% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco68.R
\name{isco68_to_isei}
\alias{isco68_to_isei}
\title{Translate ISCO68 to ISEI}
\usage{
isco68_to_isei(x)
}
\arguments{
\item{x}{A character vector of ISCO68 codes.}
}
\value{
A character vector of ISEI codes.
}
\description{
This function translates a vector of ISCO68 codes to ISEI codes using the
translation table stored in the \code{all_schemas$isco68_to_isei} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(ISEI = isco68_to_isei(isco68))

}
