% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco68.R
\name{isco68_to_siops}
\alias{isco68_to_siops}
\title{Translate ISCO68 to SIOPS}
\usage{
isco68_to_siops(x)
}
\arguments{
\item{x}{A character vector of ISCO68 codes.}
}
\value{
A character vector of SIOPS codes.
}
\description{
This function translates a vector of ISCO68 codes to SIOPS codes using the
translation table stored in the \code{all_schemas$isco68_to_siops} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(SIOPS = isco68_to_siops(isco68))

}
