% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R, R/isco68.R
\name{isco88_to_egp}
\alias{isco88_to_egp}
\alias{isco68_to_egp}
\title{Translate 4-digit ISCO88/ISCO68 to EGP}
\usage{
isco88_to_egp(x, self_employed, n_employees, label = FALSE)

isco68_to_egp(x, self_employed, n_employees, label = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees under each respondent.}

\item{label}{A logical value indicating whether to return the labels of the translated EGP codes (default is \code{FALSE}).}
}
\value{
A character vector of EGP codes.
}
\description{
This function translates a vector of 4-digit ISCO88/ISCO68 codes to EGP codes using the translation tables stored in \code{all_schema$isco88_to_egp / all_schema$isco68_to_egp}.
}
\details{
This translation was taken from the \code{iscogen} Stata package. For more details, check out the package documentation and search for \code{ISCO88/IS68 -> EGP}.
}
\examples{
library(dplyr)

# isco88
ess \%>\% transmute(
  isco88,
  egp = isco88_to_egp(isco88, self_employed, emplno, label = FALSE),
  egp_label = isco88_to_egp(isco88, self_employed, emplno, label = TRUE)
)

# isco68
ess \%>\% transmute(
  isco68,
  egp = isco68_to_egp(isco68, self_employed, emplno, label = FALSE),
  egp_label = isco68_to_egp(isco68, self_employed, emplno, label = TRUE)
)

}
