% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R, R/isco68.R
\name{isco88_to_egp11_mp}
\alias{isco88_to_egp11_mp}
\alias{isco68_to_egp11_mp}
\title{Translate 4-digit ISCO88/ISCO68 to EGP11-MP}
\usage{
isco88_to_egp11_mp(x, is_supervisor, self_employed, n_employees, label = FALSE)

isco68_to_egp11_mp(x, is_supervisor, self_employed, n_employees, label = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees under each respondent.}

\item{label}{A logical value indicating whether to return the labels of the translated EGP11-MP codes (default is \code{FALSE}).}
}
\value{
A character vector of EGP11-MP codes.
}
\description{
This function translates a vector of 4-digit ISCO88/ISCO68 codes to EGP11-MP codes using the translation tables stored in \code{all_schema$isco88_to_egp11 / all_schema$isco68_to_egp11}. After translating to EGP11 using these tables, this function reassigns managers and professionals (ISCO88/ISCO68 codes 1 and 2) to have both high/low managers and profesionals
}
\details{
EGP11-MP is a class schema similar to EGP11 but reassigns managers and professionals (ISCO88/ISCO68 codes 1 and 2) to have both high/low managers and profesionals.
}
\section{TODO: After this is corrected in the code, correct it here and in all docs.}{
This schema is a slight variation of the original EGP11 and the logic used to build this is like this:
\itemize{
\item All occupations with EGP11 digit 1 and ISCO 1-digit 0 or 1 or has subordinates, \strong{is a high manager}
\item All occupations with EGP11 digit 1 and is self-employed with more than 1 employee, \strong{is a high manager}
\item All occupations with EGP11 digit 1 and has a 1-digit ISCO higher than 1 and is either an employee or a self-employed with no subordinates, is a \strong{high professional}
\item All occupations with EGP11 digit 2 and ISCO 1-digit 0 or 1 or has subordinates, is a \strong{lower manager}
\item All occupations with EGP11 digit 2 and is self-employed with more than 1 employee, is a \strong{lower manager}
\item All occupations with EGP11 digit 2 and has a 1-digit ISCO higher than 1 and is either an employee or a self-employed with no subordinates, is a \strong{lower professional}
}

This translation was created from the Stata do file shared by Oscar Smallenbroek called "EGP-MP.do". For more info, please contact the author.
}

\examples{
library(dplyr)

# isco88
ess \%>\%
  transmute(
    isco88,
    egp11_mp = isco88_to_egp11_mp(
      isco88,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    ),
    egp11_mp_label = isco88_to_egp11_mp(
      isco88,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    )
  )

# isco68
ess \%>\%
  transmute(
    isco68,
    egp11_mp = isco68_to_egp11_mp(
      isco68,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    ),
    egp11_mp_label = isco68_to_egp11_mp(
      isco68,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    )
  )

}
