% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88_to_isco68}
\alias{isco88_to_isco68}
\title{Translate ISCO88 to ISCO68}
\usage{
isco88_to_isco68(x, label = FALSE)
}
\arguments{
\item{x}{A character vector of ISCO08 codes.}

\item{label}{A logical value indicating whether to return the labels of the
translated ISCO88 codes (default is \code{FALSE}).}
}
\value{
A character vector of ISCO88 codes.
}
\description{
This function translates a vector of ISCO08 codes to ISCO88 codes using the
translation table stored in the \code{all_schemas$isco08_to_isco88} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(ISCO88 = isco08_to_isco88(isco08, label = TRUE))
ess \%>\% mutate(ISCO88 = isco08_to_isco88(isco08, label = FALSE))

}
