% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88_to_isco88com}
\alias{isco88_to_isco88com}
\title{Translate 4-digit ISCO88 to ISCO88COM}
\usage{
isco88_to_isco88com(x, label = FALSE, factor = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}

\item{label}{A logical value indicating whether to return the labels of the translated ISCO88COM codes (default is \code{FALSE}).}

\item{factor}{A logical value indicating whether to return a factor instead of a character. The order of the labels is taken from the labels for ISCO88COM found in \code{all_labels} (default is \code{FALSE}).}
}
\value{
A character vector of ISCO88COM codes.
}
\description{
This function translates a vector of 4-digit ISCO88 codes to ISCO88COM codes using the translation table stored in \code{all_schema$isco88_to_isco88com}.
}
\details{
This translation was taken from the \code{iscogen} Stata package. For more details, check out the package documentation and search for \code{ISCO88 -> ISCO88COM}.
}
\examples{
library(dplyr)

ess \%>\%
  transmute(
    isco88,
    isco88com = isco88_to_isco88com(isco88, label = FALSE),
    isco88com_label = isco88_to_isco88com(isco88, label = TRUE)
  )

}
