% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88_to_mps}
\alias{isco88_to_mps}
\title{Translate ISCO88 to MPS}
\usage{
isco88_to_mps(x)
}
\arguments{
\item{x}{A character vector of ISCO88 codes.}
}
\value{
A character vector of MPS codes.
}
\description{
This function translates a vector of ISCO88 codes to MPS codes using the
translation table stored in the \code{all_schemas$isco88_to_mps} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(MPS = isco88_to_mps(isco88))

}
