% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88com_to_esec}
\alias{isco88com_to_esec}
\title{Translate 3-digit ISCO88COM to ESEC}
\usage{
isco88com_to_esec(
  x,
  is_supervisor,
  self_employed,
  n_employees,
  full_method = TRUE,
  label = FALSE
)
}
\arguments{
\item{x}{A character vector of 3-digit ISCO88COM codes. Even though these should be 3-digit, instead of 130, the code should be 1300, which is the 3-digit version of ISCO.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees for each individual.}

\item{full_method}{logical value indicating whether to do the translation using the full method (uses ISCO codes as well as information on number of employees, self-employed, etc..) or the simplified method (matches ISCO codes directly with ESEC codes).}

\item{label}{A logical value indicating whether to return the labels of the translated ESEC codes (default is \code{FALSE}).}
}
\value{
A character vector of ESEC codes.
}
\description{
This function translates a vector of 3-digit ISCO88COM codes to ESEC codes using the
translation table stored in the \code{all_schemas$isco88_to_esec} data frame. Note that in
\code{all_schemas$isco88_to_esec} the column \code{ESEC} refers to the simplified ESEC, which
is matched ISCO code to ESEC code directly, instead of using information on number
of employees, self-employed, etc..
}
\details{
The ESEC translation is from ISCO88COM to ESEC. This translation is borrowed from the \code{iscogen} Stata package. For more info, search for 'ISCO-88 -> ESEC' in the documentation of the \code{iscogen} package. If you have ISCO88, you can translate it to ISCO88COM using the function \code{isco88_to_isco88com} before translating to ESEC.

The codification used in the User's guide suggests that contrary to the full method, which uses whether the respondent is a supervisor, self-employed and whether the person has subordinates, the simple method matches directly the ISCO code to an ESEC code. For more info, please see page 17 of the European Socio-economic Classification (ESeC) User Guide (2006) by Rode, D. and Harrison, E.
}
\examples{
library(dplyr)

# convert to three digits
ess$isco88_three <- isco88_swap(ess$isco88, from = 4, to = 3)

# Using the full method
ess \%>\%
  transmute(
    esec_label = isco88com_to_esec(
      isco88_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    ),
    esec = isco88com_to_esec(
      isco88_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    )
  )

# Using the simple method
ess \%>\%
  transmute(
    esec_simple = isco88com_to_esec(
      isco88_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE,
      full_method = FALSE
    ),
    esec_simple_label = isco88com_to_esec(
      isco88_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE, full_method = FALSE
    )
 )

}
