% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88com_to_esec_mp}
\alias{isco88com_to_esec_mp}
\title{Translate 3-digit ISCO88COM to ESEC-MP}
\usage{
isco88com_to_esec_mp(
  x,
  is_supervisor,
  self_employed,
  n_employees,
  full_method = TRUE,
  label = FALSE
)
}
\arguments{
\item{x}{A character vector of 3-digit ISCO88COM codes. Even though these should be 3-digit, instead of 130, the code should be 1300, which is the 3-digit version of ISCO.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees for each individual.}

\item{full_method}{logical value indicating whether to do the translation using the full method (uses ISCO codes as well as information on number of employees, self-employed, etc..) or the simplified method (matches ISCO codes directly with ESEC codes).}

\item{label}{A logical value indicating whether to return the labels of the translated ESEC codes (default is \code{FALSE}).}
}
\value{
A character vector of ESEC-MP codes.
}
\description{
This function translates a vector of 3-digit ISCO88COM codes to ESEC-MP codes. ESEC-MP is a class schema similar to ESEC but reassigns managers and professionals (ISCO88COM codes 1 and 2) to have both high/low managers and profesionals. Note that since this translation first converts ISCO88COM to ESEC, this function has the possibility of using the 'Full-method' or the 'Simple method' for translating ISCO88COM to ESEC.
}
\details{
The codification used in the ESEC user's guide suggests that contrary to the full method, which uses whether the respondent is a supervisor, self-employed, employee and whether the person has subordinates, the simple method matches directly the ISCO code to an ESEC code. For more info, please see page 17 of the European Socio-economic Classification (ESeC) User Guide (2006) by Rode, D. and Harrison, E.

After translating from ISCO88COM to ESEC, the logic used to build ESEC-MP is the following:
\itemize{
\item All occupations with ESEC digit 1 and ISCO 1-digit 0 or 1 or has subordinates, \strong{is a high manager}
\item All occupations with ESEC digit 1 and is self-employed with more than 1 employee, \strong{is a high manager}
\item All occupations with ESEC digit 1 and has a 1-digit ISCO higher than 1 and is either an employee or a self-employed with no subordinates, is a \strong{high professional}
\item All occupations with ESEC digit 2 and ISCO 1-digit 0 or 1 or has subordinates, is a \strong{lower manager}
\item All occupations with ESEC digit 2 and is self-employed with more than 1 employee, is a \strong{lower manager}
\item All occupations with ESEC digit 2 and has a 1-digit ISCO higher than 1 and is either an employee or a self-employed with no subordinates, is a \strong{lower professional}
}

All other ESEC codes remaing the same.

This translation was created from the Stata do file shared by Oscar Smallenbroek called "ESEC-MP.do". For more info, please contact the author.

This function will accept 3-digit codes as 4 digits. This means that if the 3-digit code is 131 then it should be 1310. All codes should be 4 digits, even though the code is represented as 3 digits (1310, 1230, etc..)
}
\examples{

library(dplyr)

# Convert to three digits and isc08com
ess$isco88com <- isco88_to_isco88com(ess$isco88)
ess$isco88com_three <- isco88_swap(ess$isco88com, from = 4, to = 3)

# Using the full method
ess \%>\%
  transmute(
    isco88com_three,
    esec_label = isco88com_to_esec_mp(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      full_method = TRUE,
      label = TRUE
    ),
    esec = isco88com_to_esec_mp(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      full_method = TRUE,
      label = FALSE
    )
  )

# Using the simple method
ess \%>\%
  transmute(
    isco88com_three,
    esec_simple = isco88com_to_esec_mp(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      full_method = FALSE,
      label = FALSE
    ),
    esec_simple_label = isco88com_to_esec_mp(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      full_method = FALSE,
      label = TRUE
    )
  )

}
