% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_isco.R
\name{repair_isco}
\alias{repair_isco}
\title{Repair an ISCO variable}
\usage{
repair_isco(x)
}
\arguments{
\item{x}{ISCO variable to repair}
}
\value{
Repaired ISCO variable
}
\description{
ISCO variables need have certain properties. In particular, they must contain occupations of a single digit. This means that all ocupations need to be either 4 digits, 3 digits, 2 digits or 1 digit. The ISCO variable can't have 3 digits and 2 digits at the same time.
}
\details{
When reading datasets, it might happen that the ISCO column is read as a numeric column and occupations such as \code{0140} are converted to \code{140}. It is impossible for \code{DIGCLASS} to determine whether this class is actually \code{0140} or \code{1400}. what this function does is a bit of sanity check:
\itemize{
\item Checks if the provided ISCO variable is a character vector, and if not, warns the user that numeric ISCO variables may contain lost data.
\item Checks that all occupations have the same number of digits, and if not, raises an error.
}
}
\examples{

repair_isco(c("1234", "5678", "9012"))

\dontrun{
repair_isco(c("123", "5678", "012"))
}

}
