% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess}
\alias{ess}
\title{An example data frame with data from the European Social Survey, round 5.}
\format{
Data Frame
}
\description{
Just an example data frame to show how the package can be used to translate
between schemas.
}
\details{
The codebook for the columns is:
\itemize{
\item \code{isco68}: The ISCO68 class schema in 4-digits.
\item \code{isco88}: The ISCO88 class schema in 4-digits.
\item \code{isco88com}: The ISCO88COM class schema in 4-digits.
\item \code{isco08}: The ISCO08 class schema in 4-digits.
\item \code{emplno}: Number of subordinates, if there are any. If 0 employees, the number is 0 and not an \code{NA}, as it should be in your data for using \code{DIGCLASS}.
\item \code{self_employed}: A numeric vector indicating whether each individual is self-employed (1) or an employee (0).
\item \code{is_supervisor}: A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).
\item \code{control_work}: A likert-scale type question from 0 to 10 where 0 is whether an individual has no control over their work/organisation decisions and 10 is complete control over work/organization decisions. For an example, see the variable \code{iorgact} in the European Social Survey.
\item \code{control_daily}: A likert-scale type question from 1 to 4 where 1 means complete control to decide how their own daily work is/was organised and 4 means no control to decide how their own daily work is/was organised. For an example, see the variable \code{orgwrk} in the European Social Survey. Another example is recoding the variable \code{wkdcorga} from the European Social Survey such that 8-10 is 1, 5-7 is 2, 2-4 is 3 and 0-1 is 4.
\item \code{work_status}: This is the same as \code{self_employed} but has non-employed respondents coded as \code{2}. The unique values are \code{1 = self_employed}, \code{0 = employee} and \verb{2 = non employed}. This variable was constructed using the variable \code{emplrel} and \code{mainact} for the unemployed.
\item \code{main_activity}: A numeric vector where \verb{1 = respondent is working}, \verb{2 = respondent is in education}, \verb{3 = respondent is disabled }, \verb{4 = respondent has no paid work (household work, taking care of children, etc..)} and \verb{5 = respondent is retired}. This variable can be recoded using the \code{mainact} variable from the European Social Survey.
}
}
