% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R, R/isco88.R, R/isco68.R
\name{isco08_swap}
\alias{isco08_swap}
\alias{isco88_swap}
\alias{isco68_swap}
\title{Swap ISCO08/ISCO88/ISCO88 between 1, 2, 3 and 4 digit groups}
\usage{
isco08_swap(x, from, to)

isco88_swap(x, from, to)

isco68_swap(x, from, to)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO08/ISCO88/ISCO88 codes. By 4 digit it means that even though the function could be 3-digits (code 131 for example), the code should be 1310, which is the 4-digit version of ISCO.}

\item{from}{a numeric specifying the occupation digits of the input vector. Possible values are only 1, 2, 3 or 4.}

\item{to}{a numeric specifying the desired occupation digits. Possible values are only 1, 2, 3 or 4.}
}
\value{
A character vector of ISCO08/ISCO88/ISCO88 codes.
}
\description{
This function translates a vector of ISCO08/ISCO88/ISCO88 codes between different digits. For most surveys, this will be translating between the 4 digit occupations to more general groups, such as three , two and one digit groups.
}
\details{
Note that to translate using \verb{isco*_swap} you'll need to provide the \code{from} and \code{to} arguments. The first one specifies the current number of digits of the input variable. If your variable is 4-digit occupations, then \code{from} should be \code{4}. If you want to translate 4-digit occupations to 3-digits then the arguments should be \code{from = 4} and \code{to = 3}. See the argument description of \code{from} and \code{to} for all possible values. As well as examples on how this works.

This function will accept 4 digit codes as 4 digits. This means that if the 3-digit code is 131 then it should be 1310. All codes should be 4 digits, even though the code is represented as 3-digits (1310, 1320, etc..)

Note that translation can only be done from higher to smaller digits (4 to 3, 3 to 2, 3 to 1) and never the other way around (1 to 2, 2 to 3, 3 to 4).

ISCO68 might return some NAs depending on the occupation code as it does not have 4 digits for the groups 0000 and 1000. Any translation from 4 digit codes to 1 digit codes within those groups will return an NA for those major groups. See the ILO website: \url{https://www.ilo.org/public/english/bureau/stat/isco/isco68/major.htm}.
}
\examples{
library(dplyr)

# isco08
ess \%>\%
  transmute(
    isco08,
    isco08_one = isco08_swap(isco08, from = 4, to = 1),
    isco08_two = isco08_swap(isco08, from = 4, to = 2),
    isco08_three = isco08_swap(isco08, from = 4, to = 3),
    isco08_four = isco08_swap(isco08, from = 4, to = 4)
  )

# isco88
ess \%>\%
  transmute(
    isco88,
    isco88_one = isco88_swap(isco88, from = 4, to = 1),
    isco88_two = isco88_swap(isco88, from = 4, to = 2),
    isco88_three = isco88_swap(isco88, from = 4, to = 3),
    isco88_four = isco88_swap(isco88, from = 4, to = 4)
  )

# isco68
# Note that for certain four digit groups, isco68 does not have a
# major group (0000, 1000). That means that Some NAs might be present,
# such as for occupations that are between 1000 and 200. Remember to
# check well the result.
ess \%>\%
  transmute(
    isco68,
    isco68_one = isco68_swap(isco68, from = 4, to = 1),
    isco68_two = isco68_swap(isco68, from = 4, to = 2),
    isco68_three = isco68_swap(isco68, from = 4, to = 3),
    isco68_four = isco68_swap(isco68, from = 4, to = 4)
  )

}
