% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R
\name{isco08_to_esec}
\alias{isco08_to_esec}
\title{Translate 3-digit ISCO08 to ESEC}
\usage{
isco08_to_esec(x, is_supervisor, self_employed, n_employees, label = FALSE)
}
\arguments{
\item{x}{A character vector of 3-digit ISCO08 codes. Even though these should be 3-digit, instead of 130, the code should be 1300, which is the 3-digit version of ISCO.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees for each individual.}

\item{label}{A logical value indicating whether to return the labels of the translated ESEC codes (default is \code{FALSE}).}
}
\value{
A character vector of ESEC codes.
}
\description{
This function translates a vector of 3-digit ISCO08 codes to ESEC codes using the
translation table stored in the \code{all_schemas$isco08_to_esec_three} data frame.
}
\details{
This translation is borrowed from the \code{iscogen} Stata package. For more info, search for 'ISCO-08 -> ESEC' in the documentation of the \code{iscogen} package.

This function will accept 3-digit codes as 4 digits. This means that if the 3-digit code is 131 then it should be 1310. All codes should be 4 digits, even though the code is represented as 3 digits (1310, 1230, etc..)

Contrary to ISCO88COM-ESEC, ISCO08 does not have a simplified method and the translation is done from ISCO08 and not ISCO08COM.
}
\examples{
library(dplyr)

# convert to three digits
ess$isco08_three <- isco08_swap(ess$isco08, from = 4, to = 3)

ess \%>\%
  transmute(
    esec_label = isco08_to_esec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    ),
    esec = isco08_to_esec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    )
  )

}
