% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R, R/isco88.R
\name{isco08_to_esec_mp}
\alias{isco08_to_esec_mp}
\alias{isco88com_to_esec_mp}
\title{Translate 3-digit ISCO08/ISCO88COM to ESEC-MP}
\usage{
isco08_to_esec_mp(x, is_supervisor, self_employed, n_employees, label = FALSE)

isco88com_to_esec_mp(
  x,
  is_supervisor,
  self_employed,
  n_employees,
  full_method = TRUE,
  label = FALSE
)
}
\arguments{
\item{x}{A character vector of 3-digit ISCO codes. This should be the 4-digit equivalent so instead of 131, the code should be 1310, which is the 4-digit version of of the 3-digit ISCO.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees under each respondent.}

\item{label}{A logical value indicating whether to return the labels of the translated ESEC-MP codes (default is \code{FALSE}).}

\item{full_method}{a boolean on whether to apply the full method or the simple method.}
}
\value{
A character vector of ESEC-MP codes.
}
\description{
This function translates a vector of 3-digit ISCO08/ISCO88COM codes to ESEC-MP codes using the translation tables stored in \code{all_schema$isco08_to_esec / all_schema$isco88com_to_esec_three}. After translating to ESEC using these tables, this function reassigns managers and professionals (ISCO08/ISCO88COM codes 1 and 2) to have both high/low managers and profesionals
}
\details{
This function translates a vector of 3-digit ISCO08/ISCO88COM codes to ESEC-MP codes.

ESEC-MP is a class schema similar to ESEC but reassigns managers and professionals (ISCO08/ISCO88COM codes 1 and 2) to have both high/low managers and profesionals. Similarly to \code{DIGCLASS::isco88com_to_esec}, \code{isco88com_to_esec_mp} allows to translate using the simple or full method. \code{isco08_to_esec_mp} does not allow to translate using two different methods and uses the full method by default.
}
\section{TODO: After this is corrected in the code, correct it here and in all docs.}{
This schema is a slight variation of the original ESEC and the logic used to build this is like this:
\itemize{
\item All occupations with ESEC digit 1 and ISCO 1-digit 0 or 1 or has subordinates, \strong{is a high manager}
\item All occupations with ESEC digit 1 and is self-employed with more than 1 employee, \strong{is a high manager}
\item All occupations with ESEC digit 1 and has a 1-digit ISCO higher than 1 and is either an employee or a self-employed with no subordinates, is a \strong{high professional}
\item All occupations with ESEC digit 2 and ISCO 1-digit 0 or 1 or has subordinates, is a \strong{lower manager}
\item All occupations with ESEC digit 2 and is self-employed with more than 1 employee, is a \strong{lower manager}
\item All occupations with ESEC digit 2 and has a 1-digit ISCO higher than 1 and is either an employee or a self-employed with no subordinates, is a \strong{lower professional}
}

This translation was created from the Stata do file shared by Oscar Smallenbroek called "ESEC-MP.do". For more info, please contact the author.

This function will accept 3 digit codes as 4 digits. This means that if the 3-digit code is 131 then it should be 1310. All codes should be 4 digits, even though the code is represented as 3-digits (1310, 1320, etc..)
}

\examples{
library(dplyr)

# convert to three digits
ess$isco08_three <- isco08_swap(ess$isco08, from = 4, to = 3)

ess \%>\%
  transmute(
    isco08_three,
    esec = isco08_to_esec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    ),
    esec_label = isco08_to_esec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    )
  )

# Convert isco88com to three digits
ess$isco88com_three <- isco88_swap(ess$isco88com, from = 4, to = 3)

# Using the full method
ess \%>\%
  transmute(
    isco88com_three,
    esec = isco88com_to_esec_mp(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      full_method = TRUE,
      label = FALSE
    ),
    esec_label = isco88com_to_esec_mp(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      full_method = TRUE,
      label = TRUE
    )
  )

# Using the simple method. For esec_mp
# we need all variables (is_supervisor, self_employed, etc..)
# because we need to assign the manager/professionals depending
# these variables.
ess \%>\%
  transmute(
    isco88com_three,
    esec_simple = isco88com_to_esec_mp(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      full_method = FALSE,
      label = FALSE
    ),
    esec_simple_label = isco88com_to_esec_mp(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      full_method = FALSE,
      label = TRUE
    )
  )
}
