% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R, R/isco68.R
\name{isco08_to_isco88}
\alias{isco08_to_isco88}
\alias{isco68_to_isco88}
\title{Translate 4-digit ISCO08/ISCO68 to ISCO88}
\usage{
isco08_to_isco88(x, label = FALSE)

isco68_to_isco88(x, label = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}

\item{label}{A logical value indicating whether to return the labels of the translated ISCO codes (default is \code{FALSE}).}
}
\value{
A character vector of ISCO88 codes.
}
\description{
This function translates a vector of 4-digit ISCO08/ISCO68 codes to ISCO88 codes using the translation table stored in \code{all_schema$isco08_to_isco88 / all_schema$isco68_to_isco88}.
}
\details{
This translation was taken from the \code{iscogen} Stata package. For more details, check out the package documentation and search for \code{ISCO08/ISCO68 -> ISCO88}.
}
\examples{
library(dplyr)

# isco08
ess \%>\% transmute(
  isco08,
  isco88 = isco08_to_isco88(isco08, label = FALSE),
  isco88_label = isco08_to_isco88(isco08, label = TRUE)
)

# isco68
ess \%>\% transmute(
  isco68,
  isco88 = isco68_to_isco88(isco68, label = FALSE),
  isco88_label = isco68_to_isco88(isco68, label = TRUE)
)

}
