% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R, R/isco88.R, R/isco68.R
\name{isco08_to_isei}
\alias{isco08_to_isei}
\alias{isco88_to_isei}
\alias{isco68_to_isei}
\title{Translate 4-digit ISCO08/ISCO88/ISCO68 to ISEI}
\usage{
isco08_to_isei(x)

isco88_to_isei(x)

isco68_to_isei(x)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}
}
\value{
A character vector of ISEI codes.
}
\description{
This function translates a vector of 4-digit ISCO08/ISCO88/ISCO68 codes to ISEI codes using the translation tables stored in \code{all_schema$isco08_to_isei / all_schema$isco88_to_isei / all_schema$isco68_to_isei}.
}
\details{
This translation was taken from the \code{iscogen} Stata package. For more details, check out the package documentation and search for \code{ISCO08/ISCO88/ISCO88 -> ISEI}.

Since \code{ISEI} doesn't have any labels, the \code{labels} is not availabe in this function.
}
\examples{
library(dplyr)

ess \%>\%
  transmute(
    isco08,
    isco88,
    isco68,
    isei_08 = isco08_to_isei(isco08),
    isei_88 = isco88_to_isei(isco88),
    isei_68 = isco68_to_isei(isco68)
  )

}
