% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R
\name{isco08_to_isei}
\alias{isco08_to_isei}
\title{Translate ISCO08 to ISEI}
\usage{
isco08_to_isei(x)
}
\arguments{
\item{x}{A character vector of ISCO08 codes.}
}
\value{
A numeric vector of ISEI codes.
}
\description{
This function translates a vector of ISCO08 codes to ISEI codes using the
translation table stored in the \code{all_schemas$isco08_to_isei} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(ISEI = isco08_to_isei(isco08))

}
