% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R, R/isco88.R
\name{isco08_to_msec}
\alias{isco08_to_msec}
\alias{isco88com_to_msec}
\title{Translate 3-digit ISCO08/ISCO88COM to MSEC}
\usage{
isco08_to_msec(x, is_supervisor, self_employed, n_employees, label = FALSE)

isco88com_to_msec(x, is_supervisor, self_employed, n_employees, label = FALSE)
}
\arguments{
\item{x}{A character vector of 3-digit ISCO codes. This should be the 4-digit equivalent so instead of 131, the code should be 1310, which is the 4-digit version of of the 3-digit ISCO.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees under each respondent.}

\item{label}{A logical value indicating whether to return the labels of the translated MSEC codes (default is \code{FALSE}).}
}
\value{
A character vector of MSEC codes.
}
\description{
This function translates a vector of 3-digit ISCO08/ISCO88COM codes to MSEC codes using the translation tables stored in \code{all_schema$isco08_to_msec / all_schema$isco88com_to_msec}.
}
\details{
These translations were created from the document "Allocation rules of ISCO-08 and ISCO-88 (COM) 3-digit codes to ESEG-Revised" from Oscar Smallenbroek, Florian Hertel and Carlo Barone. For more info, please contact the authors. Although originally called ESEG-Revised, the class schema has been formally called MSEC.

This function will accept 3 digit codes as 4 digits. This means that if the 3-digit code is 131 then it should be 1310. All codes should be 4 digits, even though the code is represented as 3-digits (1310, 1320, etc..)
}
\examples{
library(dplyr)

# convert to three digits
ess$isco08_three <- isco08_swap(ess$isco08, from = 4, to = 3)
ess$isco88com_three <- isco88_swap(ess$isco88com, from = 4, to = 3)

# isco08
ess \%>\%
  transmute(
    isco08_three,
    msec_label = isco08_to_msec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    ),
    msec = isco08_to_msec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    )
  )

# isco88com
ess \%>\%
  transmute(
    isco88com_three,
    msec_label = isco88com_to_msec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    ),
    msec = isco88com_to_msec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    )
  )

}
