% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R
\name{isco08_to_siops}
\alias{isco08_to_siops}
\title{Translate ISCO08 to SIOPS}
\usage{
isco08_to_siops(x)
}
\arguments{
\item{x}{A character vector of ISCO08 codes.}
}
\value{
A character vector of SIOPS codes.
}
\description{
This function translates a vector of ISCO08 codes to SIOPS codes using the
translation table stored in the \code{all_schemas$isco08_to_siops} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(SIOPS = isco08_to_siops(isco08))

}
