% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco68.R
\name{isco68_swap}
\alias{isco68_swap}
\title{Swap ISCO68 between 1 (unit group), 2 (minor group) and 3 (major group) digit groups}
\usage{
isco68_swap(
  x,
  from = c("unit", "minor", "submajor", "major"),
  to = c("unit", "minor", "submajor", "major")
)
}
\arguments{
\item{x}{A character vector of ISCO68 codes.}

\item{from}{a string specifying the occupation group of the input vector. Possible values are only "major", "submajor", "minor" and "unit".}

\item{to}{a string specifying the desired occupation group for input vector. Possible values are only "major", "submajor", "minor" and "unit".}
}
\value{
A character vector of ISCO68 codes.
}
\description{
This function translates a vector of ISCO68 codes between different digits. For most surveys, this will be translating between the 1 digit occupations (this is called the \code{unit} group) to more general groups, such as two digits (minor group), three digits (called submajor group) and four digits (major groups).
}
\details{
Note that to translate using \code{isco68_swap} you'll need to provide the \code{from} and \code{to} arguments. The first one specifies the current number of digits of the input variable. If your variable is 1 digit occupations, then \code{from} should be \code{unit}. If you want to translate 1 digit occupations to three digits then the arguments should be \code{from = "unit"} and \code{to = "major"}. See the argument description of \code{from} and \code{to} for all possible values. As well as examples on how this works

Note that ISCO68 does not have 4 digits for the groups 0000 and 1000. Any translation to those major groups will return an NA. See the ILO website: \url{https://www.ilo.org/public/english/bureau/stat/isco/isco68/major.htm}.
}
\examples{
library(dplyr)

# Note that for certain four digit groups, isco68 does not have a
# major group (0000, 1000). That means that Some NAs might be present,
# such as for occupations that are between 1000 and 200. Remember to
# check well the result.
ess \%>\% mutate(
   isco68_four_digits = isco68_swap(isco68, from = "unit", to = "major"),
   isco68_three_digits = isco68_swap(isco68, from = "unit", to = "submajor"),
   isco68_two_digits = isco68_swap(isco68, from = "unit", to = "minor")
)

}
