% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88_swap}
\alias{isco88_swap}
\title{Swap ISCO88 between 1 (unit group), 2 (minor group), 3 (submajor group) and 4 (major) digit groups}
\usage{
isco88_swap(
  x,
  from = c("unit", "minor", "submajor", "major"),
  to = c("unit", "minor", "submajor", "major")
)
}
\arguments{
\item{x}{A character vector of ISCO88 codes.}

\item{from}{a string specifying the occupation group of the input vector. Possible values are only "major", "submajor", "minor" and "unit".}

\item{to}{a string specifying the desired occupation group for input vector. Possible values are only "major", "submajor", "minor" and "unit".}
}
\value{
A character vector of ISCO88 codes.
}
\description{
This function translates a vector of ISCO88 codes between different digits. For most surveys, this will be translating between the 1 digit occupations (this is called the \code{unit} group) to more general groups, such as two digits (minor group), three digits (called submajor group) and four digits (major groups).
}
\details{
Note that to translate using \code{isco88_swap} you'll need to provide the \code{from} and \code{to} arguments. The first one specifies the current number of digits of the input variable. If your variable is 1 digit occupations, then \code{from} should be \code{unit}. If you want to translate 1 digit occupations to three digits then the arguments should be \code{from = "unit"} and \code{to = "major"}. See the argument description of \code{from} and \code{to} for all possible values. As well as examples on how this works
}
\examples{
library(dplyr)

ess \%>\% mutate(
   isco88_four_digits = isco88_swap(isco88, from = "unit", to = "major"),
   isco88_three_digits = isco88_swap(isco88, from = "unit", to = "submajor"),
   isco88_two_digits = isco88_swap(isco88, from = "unit", to = "minor")
)

}
