% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R, R/isco68.R
\name{isco88_to_egp}
\alias{isco88_to_egp}
\alias{isco68_to_egp}
\title{Translate 4-digit ISCO88/ISCO68 to EGP11/EGP7/EGP5/EGP3}
\usage{
isco88_to_egp(x, self_employed, n_employees, n_classes = 11, label = FALSE)

isco68_to_egp(x, self_employed, n_employees, n_classes = 11, label = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or an employee (0).}

\item{n_employees}{A numeric vector indicating the number of employees under each respondent.}

\item{n_classes}{a numeric value indicating the number of EGP classes to obtain. Default is 11 EGP classes. The possible values are 11 classes, 7 classes, 5 classes and 3 classes. For more information, see the details section.}

\item{label}{A logical value indicating whether to return the labels of the translated EGP codes (default is \code{FALSE}).}
}
\value{
A character vector of EGP codes.
}
\description{
This function translates a vector of 4-digit ISCO88/ISCO68 codes to EGP11/EGP7/EGP5/EGP3 codes using the translation tables stored in \code{all_schema$isco88_to_egp11 / all_schema$isco68_to_egp11 / all_schema$egp11_to_egp7 / all_schema$egp11_to_egp5 / all_schema$egp11_to_egp3}.
}
\details{
This function works by first translating to EGP11 and then translating to other EGP variants, if the user has requested this through the \code{n_classes} argument.

This translation was taken from the \code{iscogen} Stata package. For more details, check out the package documentation and search for \code{ISCO88/IS68 -> EGP11}. For translations between EGP11 and EGP7/EGP5/EGP3, see the source of the \code{occupar} R package \href{https://github.com/DiogoFerrari/occupar/blob/7130d94438f1da2a4aac4731437991a8eea88436/R/occupar_occupation.R#L226-L288}{here}.

Users can see the translation used in this package in \code{all_schema$egp11_to_egp7}, \code{all_schema$egp11_to_egp5} and \code{all_schema$egp11_to_egp3}. Moreover, the labels used can be found in \code{all_labels$egp11}, \code{all_labels$egp7}, \code{all_labels$egp5} and \code{all_labels$egp3}.

For more details on this class schema, please check the references below:
\itemize{
\item Erikson, R., Goldthorpe, J.H. and Portocarero, L (1979) Intergenerational Class Mobility inThree Western European Societies: England, France and Sweden’, British Journal of Sociology, 30.
\item Erikson, R., J.H. Goldthorpe, L. Portocarero (1983) Intergenerational Class Mobility and the Convergence Thesis: England, France and Sweden. British Journal of Sociology, 34(3): 303-343.
\item Erikson R, Goldthorpe JH (1992) The Constant Flux: A Study of Class Mobility in Industrial Societies. Oxford: Clarendon Press.
\item Goldthorpe class scheme. Oxford Reference. Retrieved 19 May. 2023, from \url{https://www.oxfordreference.com/view/10.1093/oi/authority.20110803095858703}.
\item Goldthorpe JH (2007) Social class and the differentiation of employment contracts. In: Goldthorpe JH (ed.) On Sociology: Volume Two. Stanford: Stanford University Press,101–124.
}
}
\examples{
library(dplyr)

# isco88
ess \%>\% transmute(
  isco88,
  egp11 = isco88_to_egp(isco88, self_employed, emplno, label = FALSE),
  egp7 = isco88_to_egp(isco88, self_employed, emplno, n_classes = 7, label = FALSE),
  egp5 = isco88_to_egp(isco88, self_employed, emplno, n_classes = 5, label = FALSE),
  egp3 = isco88_to_egp(isco88, self_employed, emplno, n_classes = 3, label = FALSE),
)

# isco68
ess \%>\% transmute(
  isco68,
  egp11 = isco68_to_egp(isco68, self_employed, emplno, label = FALSE),
  egp7 = isco68_to_egp(isco68, self_employed, emplno, n_classes = 7, label = FALSE),
  egp5 = isco68_to_egp(isco68, self_employed, emplno, n_classes = 5, label = FALSE),
  egp3 = isco68_to_egp(isco68, self_employed, emplno, n_classes = 3, label = FALSE)
)

}
