% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88_to_egp11}
\alias{isco88_to_egp11}
\title{Translate ISCO88 to EGP11}
\usage{
isco88_to_egp11(x, self_employed, n_employees, label = FALSE)
}
\arguments{
\item{x}{A character vector of ISCO88 codes.}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or not (0).}

\item{n_employees}{A numeric vector indicating the number of employees for each individual.}

\item{label}{A logical value indicating whether to return the labels of the
translated EGP11 codes (default is \code{FALSE}).}
}
\value{
A character vector of EGP codes.
}
\description{
This function translates a vector of ISCO88 codes to EGP11 codes using the
translation table stored in the \code{all_schemas$isco88_to_egp11} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(EGP = isco88_to_egp11(isco88, self_employed, emplno, label = TRUE))
ess \%>\% mutate(EGP = isco88_to_egp11(isco88, self_employed, emplno, label = FALSE))

}
