% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R, R/isco68.R
\name{isco88_to_isco08}
\alias{isco88_to_isco08}
\alias{isco68_to_isco08}
\title{Translate 4-digit ISCO88/ISCO68 to ISCO08}
\usage{
isco88_to_isco08(x, label = FALSE, factor = FALSE)

isco68_to_isco08(x, label = FALSE, factor = FALSE)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}

\item{label}{A logical value indicating whether to return the labels of the translated ISCO08 codes (default is \code{FALSE}).}

\item{factor}{A logical value indicating whether to return a factor instead of a character. The order of the labels is taken from the labels for ISCO08 found in \code{all_labels} (default is \code{FALSE}).}
}
\value{
A character vector of ISCO08 codes.
}
\description{
This function translates a vector of 4-digit ISCO88/ISCO68 codes to ISCO08 codes using the translation tables stored in \code{all_schema$isco88_to_isco08 / all_schema$isco68_to_isco08}.
}
\details{
This translation was taken from the \code{iscogen} Stata package. For more details, check out the package documentation and search for \code{ISCO88/ISCO68 -> ISCO08}.
}
\examples{
library(dplyr)

# isco88
ess \%>\%
  transmute(
    isco88,
    isco08 = isco88_to_isco08(isco88, label = FALSE),
    isco08_label = isco88_to_isco08(isco88, label = TRUE)
  )

# isco68
ess \%>\%
  transmute(
    isco68,
    isco08 = isco68_to_isco08(isco68, label = FALSE),
    isco08_label = isco68_to_isco08(isco68, label = TRUE)
  )

}
