% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88_to_isco88com}
\alias{isco88_to_isco88com}
\title{Translate ISCO88 to ISCO88COM}
\usage{
isco88_to_isco88com(x, label = FALSE)
}
\arguments{
\item{x}{A character vector of ISCO88 codes.}

\item{label}{A logical value indicating whether to return the labels of the
translated ISCO88COM codes (default is \code{FALSE}).}
}
\value{
A character vector of ISCO88COM codes.
}
\description{
This function translates a vector of ISCO88 codes to ISCO88COM codes using the
translation table stored in the \code{all_schemas$isco88_to_isco88com} data frame.
}
\examples{
library(dplyr)

ess \%>\% mutate(ISCO88COM = isco88_to_isco88com(isco88, label = TRUE))
ess \%>\% mutate(ISCO88COM = isco88_to_isco88com(isco88, label = FALSE))

}
