% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88_to_mps}
\alias{isco88_to_mps}
\title{Translate 4-digit ISCO88 to MPS}
\usage{
isco88_to_mps(x)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes.}
}
\value{
A character vector of MPS codes.
}
\description{
This function translates a vector of 4-digit ISCO88 codes to MPS codes using the translation table stored in \code{all_schema$isco88_to_mps}.
}
\details{
This translation was taken from the \code{iscogen} Stata package. For more details, check out the package documentation and search for \code{ISCO88 -> MPS}.
}
\examples{
library(dplyr)

ess \%>\%
  transmute(
    isco88,
    mps = isco88_to_mps(isco88),
  )

}
