% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88.R
\name{isco88com_to_wright}
\alias{isco88com_to_wright}
\title{This function translates a vector of 4-digit ISCO88COM codes to the E.O Wright class schema.}
\usage{
isco88com_to_wright(
  x,
  is_supervisor,
  self_employed,
  n_employees,
  control_work,
  control_daily,
  type,
  label = FALSE
)
}
\arguments{
\item{x}{A character vector of 4-digit ISCO codes. This should be the 4-digit equivalent so instead of 131, the code should be 1310, which is the 4-digit version of of the 4-digit ISCO.}

\item{is_supervisor}{A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).}

\item{self_employed}{A numeric vector indicating whether each individual is self-employed (1) or an employee (0).}

\item{n_employees}{A numeric vector indicating the number of employees under each respondent.}

\item{control_work}{A likert-scale type question from 0 to 10 where 0 is whether an individual has no control over their work/organisation decisions and 10 is complete control over work/organization decisions. For an example, see the variable \code{iorgact} in the European Social Survey.}

\item{control_daily}{A likert-scale type question from 1 to 4 where 1 means complete control to decide how their own daily work is/was organised and 4 means no control to decide how their own daily work is/was organised. For an example, see the variable \code{orgwrk} in the European Social Survey. Another example is recoding the variable \code{wkdcorga} from the European Social Survey such that 8-10 is 1, 5-7 is 2, 2-4 is 3 and 0-1 is 4.}

\item{type}{The type of translation to make. Possible values are "simple", "decision-making" and "power-class".}

\item{label}{A logical value indicating whether to return the labels of the translated ESEC codes (default is \code{FALSE}).}
}
\description{
This function translates a vector of 4-digit ISCO88COM codes to Wright codes using the translation table stored in \code{all_schema$isco88com_to_wright}.
}
\details{
The translation implemented in this function was originally developed by Erik Olin Wright. There are three possible types of translations: the "simple" version, the "decision-making" version and the "power-class" version. This translation was implemented following the SPSS script from Håkon Leiulfsrud and Heidi Jensberg. For more information, please contact the authors.
}
\examples{
library(dplyr)


# E.O Wright - Simple translation
ess \%>\%
  transmute(
    isco88com,
    wr_simple = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "simple"
    ),
    wr_simple_label = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "simple",
      label = TRUE
    )
  )

# E.O Wright - Decision-making translation
ess \%>\%
  transmute(
    isco88com,
    wr_decision = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "decision-making"
    ),
    wr_decision_label = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "decision-making",
      label = TRUE
    )
  )

# E.O Wright - Power-class translation
ess \%>\%
  transmute(
    isco88com,
    wr_power = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "power-class"
    ),
    wr_power_label = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "power-class",
      label = TRUE
    )
  )

}
